More auth refactoring and improvements
