+++ server/src/auth/oauth_querystring.rs
@@ -1,0 +1,8 @@
+use oauth2::{AuthorizationCode, CsrfToken};
+use serde::Deserialize;
+
+#[derive(Deserialize)]
+pub struct OAuthQuerystring {
+    pub state: CsrfToken,
+    pub code: AuthorizationCode,
+}
